executor: use new file show_stats
